requirements.txt
flask
openai
python-dotenv
llama-index
pandas
tqdm